Fold back intermediate classes
