+++ it/neo4j/src/main/java/org/apache/beam/it/neo4j/DatabaseWaitOption.java
@@ -17,4 +17,5 @@
  */
 package org.apache.beam.it.neo4j;
 
-public interface DatabaseWaitOption {}
+public interface DatabaseWaitOption {
+}
