+++ it/neo4j/src/main/java/org/apache/beam/it/neo4j/Neo4jResourceManager.java
@@ -170,7 +170,7 @@ public class Neo4jResourceManager extends TestContainerResourceManager<Neo4jCont
   private void createDatabase(String databaseName, DatabaseWaitOption waitOption) {
     try (Session session =
         neo4jDriver.session(SessionConfig.builder().withDatabase("system").build())) {
-      String query = String.format("CREATE DATABASE $db %s", formatWaitOption(waitOption));
+      String query = String.format("CREATE DATABASE $db %s", DatabaseWaitOptions.asCypher(waitOption));
       session.run(query, Collections.singletonMap("db", databaseName)).consume();
     } catch (Exception e) {
       throw new Neo4jResourceManagerException(
@@ -182,7 +182,7 @@ public class Neo4jResourceManager extends TestContainerResourceManager<Neo4jCont
   void dropDatabase(String databaseName, DatabaseWaitOption waitOption) {
     try (Session session =
         neo4jDriver.session(SessionConfig.builder().withDatabase("system").build())) {
-      String query = String.format("DROP DATABASE $db %s", formatWaitOption(waitOption));
+      String query = String.format("DROP DATABASE $db %s", DatabaseWaitOptions.asCypher(waitOption));
       session.run(query, Collections.singletonMap("db", databaseName)).consume();
     } catch (Exception e) {
       throw new Neo4jResourceManagerException(
@@ -192,20 +192,6 @@ public class Neo4jResourceManager extends TestContainerResourceManager<Neo4jCont
 
   public String getAdminPassword() {
     return adminPassword;
-  }
-
-  private static String formatWaitOption(DatabaseWaitOption waitOption) {
-    if (waitOption instanceof DatabaseWaitInSeconds) {
-      return String.format("WAIT %d SECONDS", ((DatabaseWaitInSeconds) waitOption).getSeconds());
-    }
-    if (waitOption == DatabaseWait.WAIT) {
-      return "WAIT";
-    }
-    if (waitOption == DatabaseNoWait.NO_WAIT || waitOption == null) {
-      return "NOWAIT";
-    }
-    throw new Neo4jResourceManagerException(
-        String.format("Unsupported wait option type %s", waitOption.getClass()));
   }
 
   /** Builder for {@link Neo4jResourceManager}. */
